FlaskMedApp/FlaskProject/flask_app/app/requirements.txt
pandas
flask==2.2.2
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
flask-migrate
flask-injector
dependency-injector==4.35.0